Move model deviation and ase calculator to `deepmd_utils`
..., so they can benifit from multiple-backend DeepPot.
Update docs.
