+++ deepmd_utils/infer/__init__.py
@@ -2,5 +2,8 @@
 from .deep_pot import (
     DeepPot,
 )
+from .model_devi import (
+    calc_model_devi,
+)
 
-__all__ = ["DeepPot"]
+__all__ = ["DeepPot", "calc_model_devi"]
